requirements.txt
Django==4.2.7
djangorestframework==3.14.0
redis==5.0.1
pymongo==4.6.0
python-dotenv==1.0.0
drf-yasg==1.21.7
